requirements.txt
altgraph==0.17.4
macholib==1.16.3
mido==1.3.3
modulegraph==0.19.6
numpy==2.2.0
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
py2app==0.28.8
python-rtmidi==1.5.8
record3d==1.4
setuptools==70.3.0
